[CustomOp] Remove Eigen dependencies of float16 (#31669)
* remove eigen deps dof float16

* add cstdlib header

* replace stdlib header by cmath
